[TM-951] fix flaws in site state machine
