+++ app/Models/V2/Sites/Site.php
@@ -2,11 +2,13 @@
 
 namespace App\Models\V2\Sites;
 
+use App\Events\V2\General\EntityStatusChangeEvent;
 use App\Models\Framework;
 use App\Models\Traits\HasEntityResources;
+use App\Models\Traits\HasEntityStatusScopesAndTransitions;
 use App\Models\Traits\HasFrameworkKey;
 use App\Models\Traits\HasLinkedFields;
-use App\Models\Traits\HasSiteStatus;
+use App\Models\Traits\HasStatus;
 use App\Models\Traits\HasUpdateRequests;
 use App\Models\Traits\HasUuid;
 use App\Models\Traits\HasV2MediaCollections;
@@ -25,6 +27,8 @@ use App\Models\V2\TreeSpecies\TreeSpecies;
 use App\Models\V2\Workdays\Workday;
 use App\Models\V2\Workdays\WorkdayDemographic;
 use App\StateMachines\ReportStatusStateMachine;
+use App\StateMachines\SiteStatusStateMachine;
+use Asantibanez\LaravelEloquentStateMachines\Traits\HasStateMachines;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
@@ -34,6 +38,7 @@ use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Database\Eloquent\Relations\MorphMany;
 use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Str;
 use Laravel\Scout\Searchable;
 use OwenIt\Auditing\Auditable;
 use OwenIt\Auditing\Contracts\Auditable as AuditableContract;
@@ -57,8 +62,26 @@ class Site extends Model implements MediaModel, AuditableContract, EntityModel, 
     use HasFrameworkKey;
     use Auditable;
     use HasUpdateRequests;
-    use HasSiteStatus;
+    use HasStatus;
+    use HasStateMachines;
     use HasEntityResources;
+
+    use HasEntityStatusScopesAndTransitions {
+        approve as entityStatusApprove;
+        submitForApproval as entityStatusSubmitForApproval;
+    }
+
+    public static array $statuses = [
+        SiteStatusStateMachine::DRAFT => 'Draft',
+        SiteStatusStateMachine::AWAITING_APPROVAL => 'Awaiting approval',
+        SiteStatusStateMachine::NEEDS_MORE_INFORMATION => 'Needs more information',
+        SiteStatusStateMachine::RESTORATION_IN_PROGRESS => 'Restoration in progress',
+        SiteStatusStateMachine::APPROVED => 'Approved',
+    ];
+
+    public $stateMachines = [
+        'status' => SiteStatusStateMachine::class,
+    ];
 
     protected $auditInclude = [
         'status',
@@ -371,5 +394,15 @@ class Site extends Model implements MediaModel, AuditableContract, EntityModel, 
     public function getAuditableNameAttribute(): string
     {
         return $this->name;
+    }
+
+    public function dispatchStatusChangeEvent($user): void
+    {
+        EntityStatusChangeEvent::dispatch($user, $this, $this->name ?? '', '', $this->readable_status);
+    }
+
+    public function getViewLinkPath(): string
+    {
+        return '/' . Str::lower(explode_pop('\\', get_class($this))) . '/' . $this->uuid;
     }
 }
